Pagina final

+++ src/app/Guards/usuario.guard.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
+import { ActivatedRouteSnapshot, CanActivate, RouterStateSnapshot, UrlTree } from '@angular/router';
 import { Observable } from 'rxjs';
+import { Router } from '@angular/router';
 import { UserService } from '../Services/user.service';
 
 @Injectable({
@@ -8,38 +9,36 @@ import { UserService } from '../Services/user.service';
 })
 export class UsuarioGuard implements CanActivate {
 
-  respuesta:boolean = false;
   id: number = 0;
+  status: number = 0;
+  
   constructor(private usuarioService: UserService,private router:Router) { }
 
-  canActivate(route: ActivatedRouteSnapshot,state: RouterStateSnapshot):
-  Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
-    const token = localStorage.getItem('token')
+  async canActivate(route: ActivatedRouteSnapshot,state: RouterStateSnapshot): Promise<boolean | UrlTree>
+  {
+    try
+    {
+      const data: any = await this.usuarioService.revisarToken().toPromise();
+      this.status = data.status;
 
-    this.usuarioService.revisarToken()
-    .subscribe(
-      data => {
-        console.log(data);
-        this.respuesta = true;
-      },
-      error => {
-        console.log(error);
-        alert("Tienes que iniciar sesion!"); 
-        this.respuesta = false;
+      if (this.status == 0) 
+      {
+        alert("Tu cuenta está desactivada, contacta con el administrador.");
         localStorage.clear();
         this.router.navigate(['/login']);
+        return false;
+      } 
+      else 
+      {
+        return true;
       }
-    );
-
-    if(this.respuesta == true)
-    {
-      return true;
     }
-
-    else
+    catch(error)
     {
-      //alert("Tienes que iniciar sesion!"); 
-      //this.router.navigate(['/login']);
+      console.log(error);
+      alert("Tienes que iniciar sesión."); 
+      localStorage.clear();
+      this.router.navigate(['/login']);
       return false;
     }
   }
